+++ src/asset-loader.ts
@@ -4,7 +4,7 @@ import { Splat } from './splat';
 import { Env } from './env';
 
 import { startSpinner, stopSpinner } from './ui/spinner';
-import { deserializeFromSplat } from './splat-serializer';
+import { deserializeFromSSplat } from './splat-serializer';
 
 interface ModelLoadRequest {
     url?: string;
@@ -36,7 +36,7 @@ class AssetLoader {
 
         return new Promise<Model|Splat>((resolve, reject) => {
             const isPly = loadRequest.filename?.endsWith('.ply');
-            const isSplat = loadRequest.filename?.endsWith('.splat');
+            const isSSplat = loadRequest.filename?.endsWith('.splat');
 
             if (isPly) {
                 const asset = new Asset(
@@ -90,7 +90,7 @@ class AssetLoader {
 
                 registry.add(asset);
                 registry.load(asset);
-            } else if (isSplat) {
+            } else if (isSSplat) {
                 const asset = new Asset(
                     loadRequest.filename || loadRequest.url,
                     'ssplat',
@@ -103,34 +103,9 @@ class AssetLoader {
 
                 asset.on('load', () => {
                     stopSpinner();
-
-                    console.log(asset);
-                    // support loading 2d splats by adding scale_2 property with almost 0 scale
-                    const splatData = asset.resource.splatData;
-                    if (splatData.getProp('scale_0') && splatData.getProp('scale_1') && !splatData.getProp('scale_2')) {
-                        const scale2 = new Float32Array(splatData.numSplats).fill(Math.log(1e-6));
-                        splatData.addProp('scale_2', scale2);
-
-                        // place the new scale_2 property just after scale_1
-                        const props = splatData.getElement('vertex').properties;
-                        props.splice(props.findIndex((prop: any) => prop.name === 'scale_1') + 1, 0, props.splice(props.length - 1, 1)[0]);
-                    }
-
-                    // check the PLY contains minimal set of we expect
-                    const required = [
-                        'x', 'y', 'z',
-                        'scale_0', 'scale_1', 'scale_2',
-                        'rot_0', 'rot_1', 'rot_2', 'rot_3',
-                        'f_dc_0', 'f_dc_1', 'f_dc_2', 'opacity'
-                    ];
-                    const missing = required.filter(x => !splatData.getProp(x));
-                    if (missing.length > 0) {
-                        reject(`This file does not contain gaussian splatting data. The following properties are missing: ${missing.join(', ')}`);
-                    } else {
-                        resolve(new Splat(asset));
-                    }
+                    resolve(new Splat(asset));
                 });
-                
+
                 asset.on('error', (err: string) => {
                     stopSpinner();
                     reject(err);
